requirements.txt
pandas
numpy
plotly.express
streamlit
streamlit-option-menu